+++ CHANGELOG.md
@@ -6,3 +6,4 @@
     - Added Shared Token Authenticator HTTP middleware
     - Added REST and RPC controllers
     - Added predict single sample command
+    - Added predict probabilities of a single sample